+++ test-cases/syntax-1-instanced-fn/webpack.config.maker.js
@@ -38,7 +38,6 @@ function webpackConfig(index, mode) {
         },
       ],
     },
-    optimization: { minimize: false },
     output: {
       clean: true,
       filename: 'index.js',