+++ test-cases/syntax-2-static-fn-with-plugin/webpack.config.maker.js
@@ -37,7 +37,6 @@ function webpackConfig(index, mode) {
         },
       ],
     },
-    optimization: { minimize: false },
     output: {
       clean: true,
       filename: 'index.js',