+++ test-cases/syntax-4-loader-with-plugin/webpack.config.maker.js
@@ -36,7 +36,6 @@ function webpackConfig(index, mode) {
         },
       ],
     },
-    optimization: { minimize: false },
     output: {
       clean: true,
       filename: 'index.js',